Linting; added & tested quit command.
Expanded tests for begin game and reroll commands, improved pregame/ingame distinction.

+++ adventuregame/utility.py
@@ -1,21 +1,21 @@
 #!/usr/bin/python
 
 import math
+import os
 import random
 import re
-import iniconfig
 import tempfile
-import os
+
+import iniconfig
+
 
 __name__ = 'adventuregame.utility'
-
 
 # Python3's str class doesn't offer a method to test if the string constitutes
 # a float value so I rolled my own.
 _float_re = re.compile(r'^[+-]?([0-9]+\.|\.[0-9]+|[0-9]+\.[0-9]+|[0-9]+)$')
 
 isfloat = lambda strval: bool(_float_re.match(strval))
-
 
 digit_re = re.compile('^[0-9]+$')
 
@@ -35,6 +35,7 @@ lexical_number_in_1_99_re = re.compile('^('
 
 # The player can use lexical numbers (ie. 'one', 'fourteen', 'thirty') in commands and the `command_processor` needs to
 # be able to interpret them, so I wrote this utility function.
+
 def lexical_number_to_digits(lexical_number):
     if not lexical_number_in_1_99_re.match(lexical_number):
         return math.nan
@@ -80,6 +81,7 @@ def usage_verb(item_type, gerund=True):
 
 dice_expression_re = re.compile(r'([1-9]+)d([1-9][0-9]*)([-+][1-9][0-9]*)?')
 
+
 def roll_dice(dice_expr):
     match_obj = dice_expression_re.match(dice_expr)
     if not match_obj:
@@ -90,11 +92,17 @@ def roll_dice(dice_expr):
     modifier_to_roll = int(modifier_to_roll) if modifier_to_roll is not None else 0
     return sum(random.randint(1, sidedness_of_dice) for _ in range(0, number_of_dice)) + modifier_to_roll
 
+memoize_iniconfig_objs = dict()
+
 def iniconfig_obj_from_ini_text(ini_config_text):
+    if ini_config_text in memoize_iniconfig_objs:
+        return memoize_iniconfig_objs[ini_config_text]
     _, temp_ini_config_file = tempfile.mkstemp(suffix='.ini')
     temp_ini_config_fh = open(temp_ini_config_file, 'w')
     temp_ini_config_fh.write(ini_config_text)
     temp_ini_config_fh.close()
+    del temp_ini_config_fh
     ini_config_obj = iniconfig.IniConfig(temp_ini_config_file)
     os.remove(temp_ini_config_file)
+    memoize_iniconfig_objs[ini_config_text] = ini_config_obj
     return ini_config_obj
